fix warning

+++ shopAPI/shopAPI/Data/Repositories/ProductRepository.cs
@@ -1,38 +1,45 @@
-﻿using Microsoft.EntityFrameworkCore;
-using ShopApi.Data.Repositories.Interfaces;
-using ShopApi.Models.Entities;
-
-
-namespace ShopApi.Data.Repositories
+﻿namespace ShopApi.Data.Repositories
 {
+    using ShopApi.Data.Repositories.Interfaces;
+    using ShopApi.Models.Entities;
+
     public class ProductRepository : IProductRepository
     {
         private ShopContext _context;
-    
+
         public ProductRepository(ShopContext context)
         {
             _context = context;
         }
+
         public List<Product> GetAll()
         {
             return _context.Products.ToList();
         }
+
         public void AddProduct(Product product)
         {
             _context.Products.Add(product);
             _context.SaveChanges();
         }
 
-        public void  DeleteProduct(Product product)
+        public bool DeleteProduct(Product product)
         {
+            if (_context.OrderItems.Any(oi => oi.ProductId == product.ProductId))
+            {
+                return false;
+            }
+
             _context.Products.Remove(product);
             _context.SaveChanges();
+            return true;
         }
 
         public Product? GetProductById(int id)
         {
             return _context.Products.Find(id);
         }
+
         public void UpdateProduct(Product product)
         {
             _context.Products.Update(product);
@@ -42,7 +49,5 @@ namespace ShopApi.Data.Repositories
         {
             return _context.Products.Where(p => p.Name.Contains(name)).ToList();
         }
-
-       
     }
 }
